requirements.txt
psycopg2
django
django-extensions
https://github.com/darklow/django-suit/tarball/develop
django-admin-tools
django-messages
django-debug-toolbar
django-recaptcha
django-form-utils
django-bootstrap-form
django-tinymce
gunicorn
south
sorl-thumbnail
simplejson
psutil
xlrd
beautifulsoup
git+git://github.com/django-mptt/django-mptt.git
pytils
unidecode
